test: split workspace onboarding workflows
