--- tests/Feature/Onboarding/CreateTeamOnboardingTest.php
@@ -1,1216 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use App\Actions\Billing\StartProTrial;
-use App\Actions\Jetstream\CreateTeam as CreateTeamAction;
-use App\Enums\OnboardingReferralSource;
-use App\Enums\OnboardingUseCase;
-use App\Enums\Plan;
-use App\Features\Billing as BillingFeature;
-use App\Features\OnboardSeed;
-use App\Filament\Pages\CreateTeam;
-use App\Filament\Pages\Dashboard;
-use App\Listeners\CreateTeamCustomFields;
-use App\Models\Company;
-use App\Models\CustomField;
-use App\Models\CustomFieldValue;
-use App\Models\Note;
-use App\Models\Opportunity;
-use App\Models\People;
-use App\Models\Task;
-use App\Models\Team;
-use App\Models\User;
-use App\Policies\TeamPolicy;
-use Filament\Actions\Testing\TestAction;
-use Filament\Notifications\Notification;
-use Illuminate\Support\Facades\Mail;
-use Illuminate\Validation\ValidationException;
-use Laravel\Pennant\Feature;
-use Relaticle\Chat\Models\AiCreditBalance;
-use Relaticle\OnboardSeed\OnboardSeedManager;
-use Symfony\Component\Mailer\Envelope;
-use Symfony\Component\Mailer\Exception\TransportException;
-use Symfony\Component\Mailer\SentMessage;
-use Symfony\Component\Mailer\Transport\TransportInterface;
-use Symfony\Component\Mime\RawMessage;
-
-mutates(CreateTeam::class, CreateTeamAction::class, StartProTrial::class, OnboardSeedManager::class, CreateTeamCustomFields::class, TeamPolicy::class);
-
-// This file is the coverage for demo seeding itself, so it opts back into the
-// feature that TestCase switches off for the rest of the suite.
-beforeEach(function (): void {
-    Feature::define(OnboardSeed::class, true);
-});
-
-it('renders the create team page with wizard for teamless users', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->assertSuccessful()
-        ->assertSee('Create your workspace');
-});
-
-it('resolves every wizard heading from translations', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    // Every step's placeholders are in the DOM at once, so one render covers all four.
-    // A mistyped key would surface here as the raw dotted key instead of the copy.
-    livewire(CreateTeam::class)
-        ->assertSuccessful()
-        ->assertSee(__('filament/pages/teams.create_team.headings.workspace'))
-        ->assertSee(__('filament/pages/teams.create_team.headings.attribution'))
-        ->assertSee(__('filament/pages/teams.create_team.headings.attribution_description'))
-        ->assertSee(__('filament/pages/teams.create_team.headings.use_case'))
-        ->assertSee(__('filament/pages/teams.create_team.headings.use_case_description'))
-        ->assertSee(__('filament/pages/teams.create_team.headings.use_case_hint'))
-        ->assertSee(__('filament/pages/teams.create_team.headings.invite'))
-        ->assertSee(__('filament/pages/teams.create_team.headings.invite_description'))
-        ->assertSee(__('filament/pages/teams.create_team.headings.invite_subheading'))
-        ->assertDontSee('filament/pages/teams.create_team.headings');
-});
-
-it('renders wizard for users who already have a team', function (): void {
-    $user = User::factory()->withPersonalTeam()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->assertSuccessful()
-        ->assertSee('Create your workspace');
-});
-
-it('offers a way back to the current workspace for users who already have one', function (): void {
-    $user = User::factory()->withPersonalTeam()->create();
-
-    $this->actingAs($user);
-
-    $component = livewire(CreateTeam::class);
-
-    expect($component->instance()->getCancelUrl())
-        ->toBe(Dashboard::getUrl(['tenant' => $user->currentTeam]));
-
-    $component->assertSee(__('filament/pages/teams.create_team.actions.cancel'));
-});
-
-it('offers no way back for teamless users, who have nowhere to go', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    $component = livewire(CreateTeam::class);
-
-    expect($component->instance()->getCancelUrl())->toBeNull();
-
-    $component->assertDontSee(__('filament/pages/teams.create_team.actions.cancel'));
-});
-
-it('creates a team with onboarding fields', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'name' => 'Acme Corp',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = Team::query()->where('name', 'Acme Corp')->first();
-
-    expect($team)->not->toBeNull()
-        ->and($team->slug)->toBe('acme-corp')
-        ->and($team->onboarding_use_case)->toBe(OnboardingUseCase::Sales);
-});
-
-it('automatically starts one 14-day Cloud Pro trial after hosted onboarding', function (): void {
-    Feature::define(BillingFeature::class, true);
-
-    $user = User::factory()->create();
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'name' => 'Trial Workspace',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    /** @var Team $team */
-    $team = $user->refresh()->currentTeam;
-    $balance = AiCreditBalance::query()->where('team_id', $team->getKey())->sole();
-
-    expect($team->plan)->toBe(Plan::Pro)
-        ->and($team->onGenericTrial())->toBeTrue()
-        ->and($team->trial_ends_at?->isSameDay(now()->addDays(14)))->toBeTrue()
-        ->and($team->pro_trial_used_at)->not->toBeNull()
-        ->and($balance->credits_remaining)->toBe(Plan::Pro->credits());
-});
-
-it('starts a fresh trial for each additional hosted workspace', function (): void {
-    Feature::define(BillingFeature::class, true);
-
-    $user = User::factory()->withPersonalTeam()->create();
-    $user->currentTeam?->forceFill(['pro_trial_used_at' => now()])->save();
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'name' => 'Second Workspace',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    /** @var Team $team */
-    $team = $user->refresh()->currentTeam;
-
-    expect($team->plan)->toBe(Plan::Pro)
-        ->and($team->onGenericTrial())->toBeTrue()
-        ->and($team->pro_trial_used_at)->not->toBeNull()
-        ->and($team->hosted_free_grandfathered_at)->toBeNull();
-});
-
-it('creates a team with a custom slug', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'name' => 'Acme Corp',
-            'slug' => 'my-workspace',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = Team::query()->where('name', 'Acme Corp')->first();
-
-    expect($team)->not->toBeNull()
-        ->and($team->slug)->toBe('my-workspace');
-});
-
-it('validates slug format', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'name' => 'Acme Corp',
-            'slug' => 'INVALID SLUG!!',
-        ])
-        ->call('register')
-        ->assertHasFormErrors(['slug']);
-});
-
-it('validates slug uniqueness', function (): void {
-    $existingUser = User::factory()->create();
-    Team::factory()->create(['slug' => 'taken-slug', 'user_id' => $existingUser->id]);
-
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'name' => 'Acme Corp',
-            'slug' => 'taken-slug',
-        ])
-        ->call('register')
-        ->assertHasFormErrors(['slug']);
-});
-
-it('requires a team name', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'name' => '',
-        ])
-        ->call('register')
-        ->assertHasFormErrors(['name']);
-});
-
-it('marks first team as personal team', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'name' => 'My First Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = $user->fresh()->ownedTeams->first();
-
-    expect($team->personal_team)->toBeTrue();
-});
-
-it('marks subsequent teams as non-personal', function (): void {
-    $user = User::factory()->withPersonalTeam()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'name' => 'Second Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $secondTeam = $user->fresh()->ownedTeams()->where('name', 'Second Team')->first();
-
-    expect($secondTeam->personal_team)->toBeFalse();
-});
-
-it('redirects first team to dashboard with notification', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'name' => 'Redirect Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors()
-        ->assertNotified('Workspace created')
-        ->assertRedirect(Dashboard::getUrl(['tenant' => $user->fresh()->currentTeam]));
-});
-
-it('redirects subsequent teams to dashboard', function (): void {
-    $user = User::factory()->withPersonalTeam()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'name' => 'Second Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors()
-        ->assertRedirect(Dashboard::getUrl(['tenant' => $user->fresh()->currentTeam]));
-});
-
-it('seeds sales demo data for sales use case', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'name' => 'Sales Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = $user->fresh()->personalTeam();
-
-    expect($team)->not->toBeNull();
-
-    $companies = Company::where('team_id', $team->id)->pluck('name')->sort()->values();
-
-    expect($companies)->toHaveCount(4)
-        ->and($companies->all())->toBe(['Airbnb', 'Apple', 'Figma', 'Notion']);
-});
-
-it('seeds recruiting demo data for recruiting use case', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Recruiting->value,
-            'onboarding_context' => ['applications'],
-            'name' => 'Hiring Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = $user->fresh()->personalTeam();
-
-    $companies = Company::where('team_id', $team->id)->pluck('name')->sort()->values();
-    $people = People::where('team_id', $team->id)->pluck('name')->sort()->values();
-
-    expect($companies)->toHaveCount(4)
-        ->and($companies->all())->toBe(['Linear', 'Stripe', 'Supabase', 'Vercel'])
-        ->and($people)->toHaveCount(4);
-});
-
-it('seeds marketing demo data for marketing use case', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Marketing->value,
-            'onboarding_context' => ['content'],
-            'name' => 'Marketing Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = $user->fresh()->personalTeam();
-
-    $companies = Company::where('team_id', $team->id)->pluck('name')->sort()->values();
-
-    expect($companies)->toHaveCount(4)
-        ->and($companies->all())->toBe(['Canva', 'Clearbit', 'HubSpot', 'Mailchimp']);
-});
-
-it('seeds general demo data for other use case', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'name' => 'General Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = $user->fresh()->personalTeam();
-
-    $companies = Company::where('team_id', $team->id)->pluck('name')->sort()->values();
-
-    expect($companies)->toHaveCount(4)
-        ->and($companies->all())->toBe(['Atlas Design Studio', 'Coastal Media', 'Horizon Labs', 'Summit Group']);
-});
-
-it('seeds fundraising demo data for fundraising use case', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Fundraising->value,
-            'onboarding_context' => ['early_stage'],
-            'name' => 'Fundraising Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = $user->fresh()->personalTeam();
-
-    $companies = Company::where('team_id', $team->id)->pluck('name')->sort()->values();
-
-    expect($companies)->toHaveCount(4)
-        ->and($companies->all())->toBe(['Andreessen Horowitz', 'Benchmark', 'Greylock Partners', 'Sequoia Capital']);
-});
-
-it('creates all custom fields for the first team', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'name' => 'Custom Fields Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = $user->fresh()->personalTeam();
-
-    $fields = CustomField::withoutGlobalScopes()
-        ->where('tenant_id', $team->id)
-        ->get()
-        ->groupBy('entity_type');
-
-    expect($fields->get('company'))->toHaveCount(3)
-        ->and($fields->get('people'))->toHaveCount(4)
-        ->and($fields->get('opportunity'))->toHaveCount(3)
-        ->and($fields->get('task'))->toHaveCount(4)
-        ->and($fields->get('note'))->toHaveCount(1);
-});
-
-it('seeds people linked to their correct companies for sales', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'name' => 'Link Test Team',
-        ])
-        ->call('register');
-
-    $team = $user->fresh()->personalTeam();
-    $companies = Company::where('team_id', $team->id)->pluck('id', 'name');
-    $people = People::where('team_id', $team->id)->get();
-
-    $expectedMapping = [
-        'Tim Cook' => 'Apple',
-        'Brian Chesky' => 'Airbnb',
-        'Dylan Field' => 'Figma',
-        'Ivan Zhao' => 'Notion',
-    ];
-
-    foreach ($people as $person) {
-        $expectedCompany = $expectedMapping[$person->name];
-        expect($person->company_id)->toBe($companies[$expectedCompany]);
-    }
-});
-
-it('seeds tasks and opportunities with board positions', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'name' => 'Board Test Team',
-        ])
-        ->call('register');
-
-    $team = $user->fresh()->personalTeam();
-
-    $tasks = Task::where('team_id', $team->id)->get();
-    $taskPositions = $tasks->pluck('order_column');
-    expect($tasks)->toHaveCount(4)
-        ->and($taskPositions->every(fn ($v) => $v !== null))->toBeTrue()
-        ->and($taskPositions->unique())->toHaveCount(4);
-
-    $opportunities = Opportunity::where('team_id', $team->id)->get();
-    $opportunityPositions = $opportunities->pluck('order_column');
-    expect($opportunities)->toHaveCount(4)
-        ->and($opportunityPositions->every(fn ($v) => $v !== null))->toBeTrue()
-        ->and($opportunityPositions->unique())->toHaveCount(4);
-});
-
-it('seeds custom field values correctly for sales', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'name' => 'Values Test Team',
-        ])
-        ->call('register');
-
-    $team = $user->fresh()->personalTeam();
-
-    $apple = Company::where('team_id', $team->id)->where('name', 'Apple')->first();
-    $companyFields = CustomField::withoutGlobalScopes()
-        ->where('tenant_id', $team->id)
-        ->forEntity(Company::class)
-        ->pluck('id', 'code');
-
-    $appleValues = CustomFieldValue::withoutGlobalScopes()
-        ->where('entity_id', $apple->id)
-        ->where('entity_type', $apple->getMorphClass())
-        ->get()
-        ->keyBy('custom_field_id');
-
-    expect($appleValues[$companyFields['domains']]->json_value)->toContain('www.apple.com')
-        ->and($appleValues[$companyFields['icp']]->boolean_value)->toBeTrue()
-        ->and($appleValues[$companyFields['linkedin']]->json_value)->toContain('www.linkedin.com/company/apple');
-});
-
-it('subsequent teams still require use case selection', function (): void {
-    $user = User::factory()->withPersonalTeam()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Second Team',
-            'slug' => 'second-team',
-        ])
-        ->call('register')
-        ->assertHasFormErrors(['onboarding_use_case' => 'required']);
-});
-
-it('subsequent teams can skip optional referral source', function (): void {
-    $user = User::factory()->withPersonalTeam()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Second Team',
-            'slug' => 'second-team',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = $user->fresh()->ownedTeams()->where('name', 'Second Team')->first();
-
-    expect($team)->not->toBeNull()
-        ->and($team->onboarding_referral_source)->toBeNull();
-});
-
-it('provides sub-options for each use case', function (): void {
-    expect(OnboardingUseCase::Sales->getSubOptions())->toHaveCount(7)
-        ->and(OnboardingUseCase::CustomerSuccess->getSubOptions())->toHaveCount(5)
-        ->and(OnboardingUseCase::Recruiting->getSubOptions())->toHaveCount(2)
-        ->and(OnboardingUseCase::Marketing->getSubOptions())->toHaveCount(4)
-        ->and(OnboardingUseCase::Fundraising->getSubOptions())->toHaveCount(3)
-        ->and(OnboardingUseCase::Investing->getSubOptions())->toHaveCount(3)
-        ->and(OnboardingUseCase::Other->getSubOptions())->toBeEmpty();
-});
-
-it('stores referral source', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'onboarding_referral_source' => OnboardingReferralSource::Google->value,
-            'name' => 'Referral Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = Team::query()->where('name', 'Referral Team')->first();
-
-    expect($team->onboarding_referral_source)->toBe(OnboardingReferralSource::Google);
-});
-
-it('stores onboarding context', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led', 'enterprise'],
-            'name' => 'Context Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = Team::query()->where('name', 'Context Team')->first();
-
-    expect($team->onboarding_context)->toBe(['product_led', 'enterprise']);
-});
-
-it('sends team invitations when invite emails are provided', function (): void {
-    Mail::fake();
-
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Invite Test Team',
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'invites' => [
-                ['email' => 'alice@example.com', 'role' => 'editor'],
-                ['email' => 'bob@example.com', 'role' => 'editor'],
-            ],
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = Team::query()->where('name', 'Invite Test Team')->first();
-
-    expect($team->teamInvitations)->toHaveCount(2)
-        ->and($team->teamInvitations->pluck('email')->sort()->values()->all())
-        ->toBe(['alice@example.com', 'bob@example.com']);
-});
-
-it('sends invitations with correct roles', function (): void {
-    Mail::fake();
-
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Role Test Team',
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'invites' => [
-                ['email' => 'member@example.com', 'role' => 'editor'],
-                ['email' => 'admin@example.com', 'role' => 'admin'],
-            ],
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = Team::query()->where('name', 'Role Test Team')->first();
-    $invitations = $team->teamInvitations->sortBy('email')->values();
-
-    expect($invitations)->toHaveCount(2)
-        ->and($invitations[0]->email)->toBe('admin@example.com')
-        ->and($invitations[0]->role)->toBe('admin')
-        ->and($invitations[1]->email)->toBe('member@example.com')
-        ->and($invitations[1]->role)->toBe('editor');
-});
-
-it('sends only valid invitations when some emails are empty', function (): void {
-    Mail::fake();
-
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Partial Invite Team',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'invites' => [
-                ['email' => 'alice@example.com', 'role' => 'editor'],
-                ['email' => '', 'role' => 'editor'],
-            ],
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = Team::query()->where('name', 'Partial Invite Team')->first();
-
-    expect($team->teamInvitations)->toHaveCount(1)
-        ->and($team->teamInvitations->first()->email)->toBe('alice@example.com');
-});
-
-it('creates team without invitations when no emails are provided', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'No Invite Team',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = Team::query()->where('name', 'No Invite Team')->first();
-
-    expect($team->teamInvitations)->toBeEmpty();
-});
-
-it('maps use case to correct fixture set', function (): void {
-    expect(OnboardingUseCase::Sales->getFixtureSet())->toBe('sales')
-        ->and(OnboardingUseCase::CustomerSuccess->getFixtureSet())->toBe('sales')
-        ->and(OnboardingUseCase::Recruiting->getFixtureSet())->toBe('recruiting')
-        ->and(OnboardingUseCase::Marketing->getFixtureSet())->toBe('marketing')
-        ->and(OnboardingUseCase::Fundraising->getFixtureSet())->toBe('fundraising')
-        ->and(OnboardingUseCase::Investing->getFixtureSet())->toBe('fundraising')
-        ->and(OnboardingUseCase::Other->getFixtureSet())->toBe('general');
-});
-
-it('seeds all entity types for each fixture set', function (OnboardingUseCase $useCase, ?array $context): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    $formData = [
-        'onboarding_use_case' => $useCase->value,
-        'name' => "Team {$useCase->value}",
-    ];
-
-    if ($context !== null) {
-        $formData['onboarding_context'] = $context;
-    }
-
-    livewire(CreateTeam::class)
-        ->fillForm($formData)
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = $user->fresh()->personalTeam();
-
-    expect(Company::where('team_id', $team->id)->count())->toBe(4)
-        ->and(People::where('team_id', $team->id)->count())->toBe(4)
-        ->and(Opportunity::where('team_id', $team->id)->count())->toBe(4)
-        ->and(Task::where('team_id', $team->id)->count())->toBe(4)
-        ->and(Note::where('team_id', $team->id)->count())->toBe(5);
-})->with([
-    'sales' => [OnboardingUseCase::Sales, ['product_led']],
-    'recruiting' => [OnboardingUseCase::Recruiting, ['applications']],
-    'marketing' => [OnboardingUseCase::Marketing, ['content']],
-    'customer_success' => [OnboardingUseCase::CustomerSuccess, ['low_touch']],
-    'fundraising' => [OnboardingUseCase::Fundraising, ['early_stage']],
-    'investing' => [OnboardingUseCase::Investing, ['early_stage']],
-    'other' => [OnboardingUseCase::Other, null],
-]);
-
-it('rejects empty onboarding context for use cases that have sub-options', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    expect(fn () => resolve(CreateTeamAction::class)->create($user, [
-        'name' => 'Tampered Team',
-        'slug' => 'tampered-team',
-        'onboarding_use_case' => OnboardingUseCase::Sales->value,
-        'onboarding_context' => [],
-    ]))->toThrow(ValidationException::class);
-
-    expect(Team::where('slug', 'tampered-team')->exists())->toBeFalse();
-});
-
-it('rejects unknown onboarding context values for the chosen use case', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    expect(fn () => resolve(CreateTeamAction::class)->create($user, [
-        'name' => 'Tampered Team',
-        'slug' => 'tampered-team',
-        'onboarding_use_case' => OnboardingUseCase::Sales->value,
-        'onboarding_context' => ['not_a_real_option'],
-    ]))->toThrow(ValidationException::class);
-
-    expect(Team::where('slug', 'tampered-team')->exists())->toBeFalse();
-});
-
-it('allows empty onboarding context for use cases without sub-options', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    $team = resolve(CreateTeamAction::class)->create($user, [
-        'name' => 'Other Use Case',
-        'slug' => 'other-use-case',
-        'onboarding_use_case' => OnboardingUseCase::Other->value,
-    ]);
-
-    expect($team->slug)->toBe('other-use-case')
-        ->and($team->onboarding_use_case)->toBe(OnboardingUseCase::Other);
-});
-
-it('generates a fallback handle for names that transliterate to nothing', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    $state = livewire(CreateTeam::class)
-        ->fillForm(['name' => '株式会社テスト'])
-        ->get('data');
-
-    expect($state['slug'] ?? null)->toBeString()
-        ->not->toBe('')
-        ->toMatch(Team::SLUG_REGEX);
-});
-
-it('warns about invites the form accepted but that are not deliverable', function (): void {
-    Mail::fake();
-
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Invite Reporting',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'invites' => [
-                ['email' => 'user@example', 'role' => 'editor'],
-                ['email' => 'real.person@gmail.com', 'role' => 'editor'],
-            ],
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors()
-        ->assertNotified(__('filament/pages/teams.create_team.notifications.some_invites_failed.title'));
-
-    $team = Team::query()->where('name', 'Invite Reporting')->firstOrFail();
-
-    expect($team->teamInvitations()->pluck('email')->all())->toBe(['real.person@gmail.com']);
-});
-
-it('finishes onboarding when the mail transport is down', function (): void {
-    config()->set('mail.mailers.failing', ['transport' => 'failing']);
-    config()->set('mail.default', 'failing');
-
-    Mail::extend('failing', fn (array $config): TransportInterface => new class implements TransportInterface
-    {
-        public function send(RawMessage $message, ?Envelope $envelope = null): ?SentMessage
-        {
-            throw new TransportException('Connection could not be established.');
-        }
-
-        public function __toString(): string
-        {
-            return 'failing';
-        }
-    });
-
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Mail Transport Down',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'invites' => [
-                ['email' => 'first.person@gmail.com', 'role' => 'editor'],
-                ['email' => 'second.person@gmail.com', 'role' => 'editor'],
-            ],
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors()
-        ->assertNotified(
-            Notification::make()
-                ->title(__('filament/pages/teams.create_team.notifications.some_invites_failed.title'))
-                ->body(
-                    'first.person@gmail.com: '.__('filament/pages/teams.create_team.notifications.some_invites_failed.send_failed')
-                    ."\n".'second.person@gmail.com: '.__('filament/pages/teams.create_team.notifications.some_invites_failed.send_skipped')
-                )
-                ->warning()
-        )
-        ->assertRedirect(Dashboard::getUrl(['tenant' => $user->fresh()->currentTeam]));
-
-    $team = Team::query()->where('name', 'Mail Transport Down')->firstOrFail();
-
-    // The row is written before the send, so the first invitation survives and the owner
-    // can resend it. The second address is never attempted: one dead connection is enough
-    // to know the rest would only wait out the same socket timeout. The two addresses
-    // therefore need different advice, which is what the notification body asserts above.
-    expect($team->teamInvitations()->pluck('email')->all())->toBe(['first.person@gmail.com']);
-});
-
-it('assigns seeded demo tasks to the workspace owner so the dashboard is not empty', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'onboarding_use_case' => OnboardingUseCase::Sales->value,
-            'onboarding_context' => ['product_led'],
-            'name' => 'Assigned Tasks Team',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = $user->fresh()->personalTeam();
-
-    $tasks = Task::where('team_id', $team->id)->get();
-
-    expect($tasks)->not->toBeEmpty();
-
-    $tasks->each(function (Task $task) use ($user): void {
-        expect($task->assignees()->whereKey($user->getKey())->exists())->toBeTrue();
-    });
-});
-
-it('does not send invites when the user skips the invite step', function (): void {
-    Mail::fake();
-
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Skipped Invites',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'invites' => [
-                ['email' => 'not.invited@gmail.com', 'role' => 'editor'],
-            ],
-        ])
-        ->call('skipInvites')
-        ->assertHasNoFormErrors();
-
-    $team = Team::query()->where('name', 'Skipped Invites')->firstOrFail();
-
-    expect($team->teamInvitations()->count())->toBe(0);
-});
-
-it('still sends invites when the user confirms them', function (): void {
-    Mail::fake();
-
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Confirmed Invites',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-            'invites' => [
-                ['email' => 'really.invited@gmail.com', 'role' => 'editor'],
-            ],
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    $team = Team::query()->where('name', 'Confirmed Invites')->firstOrFail();
-
-    expect($team->teamInvitations()->pluck('email')->all())->toBe(['really.invited@gmail.com']);
-});
-
-it('shows a step indicator and a back affordance in the wizard', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->assertSuccessful()
-        ->assertSee(__('filament/pages/teams.create_team.actions.back'))
-        ->assertSee('Step :current of :total');
-});
-
-it('relabels cancel once copying the invite link has created the workspace', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    $component = livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Pre Created',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-        ]);
-
-    expect($component->instance()->getCancelLabel())
-        ->toBe(__('filament/pages/teams.create_team.actions.cancel'));
-
-    $component->callAction(TestAction::make('copyInviteLink')->schemaComponent());
-
-    $team = Team::query()->where('name', 'Pre Created')->firstOrFail();
-
-    expect($component->instance()->getCancelLabel())
-        ->toBe(__('filament/pages/teams.create_team.actions.go_to_workspace'))
-        ->and($component->instance()->getCancelUrl())
-        ->toBe(Dashboard::getUrl(['tenant' => $team]));
-});
-
-it('reconciles name and slug edited after the invite link pre-created the workspace', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    $component = livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'First Name',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-        ]);
-
-    $component->callAction(TestAction::make('copyInviteLink')->schemaComponent());
-
-    $team = Team::query()->where('name', 'First Name')->firstOrFail();
-
-    // The Back button makes this reachable: the user returns to step 1 and renames the
-    // workspace after it has already been created.
-    $component
-        ->fillForm([
-            'name' => 'Renamed Later',
-            'slug' => 'renamed-later',
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    expect($team->fresh())
-        ->name->toBe('Renamed Later')
-        ->slug->toBe('renamed-later')
-        ->and(Team::query()->count())->toBe(1);
-});
-
-it('labels the submit button for what it will actually do', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    $component = livewire(CreateTeam::class)
-        ->fillForm(['name' => 'Label Check', 'onboarding_use_case' => OnboardingUseCase::Other->value]);
-
-    $component
-        ->assertSee(__('filament/pages/teams.create_team.actions.get_started'))
-        ->assertDontSee(__('filament/pages/teams.create_team.actions.send_invites'));
-
-    $component->fillForm([
-        'invites' => [['email' => 'someone@gmail.com', 'role' => 'editor']],
-    ]);
-
-    $component
-        ->assertSee(__('filament/pages/teams.create_team.actions.send_invites'))
-        ->assertDontSee(__('filament/pages/teams.create_team.actions.get_started'));
-});
-
-it('allows a fourth workspace under the default cap', function (): void {
-    $user = User::factory()->create();
-
-    Team::factory()->count(3)->create(['user_id' => $user->id]);
-
-    $this->actingAs($user);
-
-    $this->get(route('filament.app.tenant.registration'))->assertSuccessful();
-});
-
-it('explains the workspace limit instead of returning a bare 404', function (): void {
-    // A low cap keeps the fixture small; the point is the behavior at the cap,
-    // whatever the configured number happens to be.
-    config()->set('relaticle.workspaces.max_owned_per_user', 3);
-
-    $user = User::factory()->create();
-
-    Team::factory()->count(3)->create(['user_id' => $user->id]);
-
-    $this->actingAs($user);
-
-    expect($user->refresh()->ownedTeams()->count())->toBe(3);
-
-    $this->get(route('filament.app.tenant.registration'))
-        ->assertRedirect()
-        ->assertSessionHas('filament.notifications');
-});
-
-it('lets a user finish a wizard run whose workspace pushed them to the limit', function (): void {
-    config()->set('relaticle.workspaces.max_owned_per_user', 3);
-
-    $user = User::factory()->create();
-
-    // Two existing workspaces: creating this one takes them to the cap of three.
-    Team::factory()->count(2)->create(['user_id' => $user->id]);
-
-    $this->actingAs($user);
-
-    $component = livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Third Workspace',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-        ]);
-
-    // Pre-creates the team, which takes the user to the cap.
-    $component->callAction(TestAction::make('copyInviteLink')->schemaComponent());
-
-    expect($user->refresh()->ownedTeams()->count())->toBe(3);
-
-    // Without the in-flight exemption Filament would 404 this request, and Livewire
-    // would swallow it: the form would simply stop responding.
-    $component
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    expect(Team::query()->where('name', 'Third Workspace')->count())->toBe(1)
-        ->and($user->refresh()->ownedTeams()->count())->toBe(3);
-});
-
-it('still refuses a brand new wizard once the user is at the limit', function (): void {
-    config()->set('relaticle.workspaces.max_owned_per_user', 3);
-
-    $user = User::factory()->create();
-
-    Team::factory()->count(3)->create(['user_id' => $user->id]);
-
-    $this->actingAs($user);
-
-    // A stale in-flight marker must not survive into a fresh visit.
-    session()->put('onboarding.completing_workspace', 'stale');
-
-    $this->get(route('filament.app.tenant.registration'))->assertRedirect();
-
-    expect($user->refresh()->ownedTeams()->count())->toBe(3);
-});
-
-/**
- * The teams table already records that a workspace exists. What only the
- * client-side event carries is the referrer still on the session, so a channel
- * can be credited with an activated workspace and not just a signup.
- *
- * Flagged in afterRegister() so it fires once per finished wizard whichever
- * path created the row, and because getRedirectUrl() sends the user to the
- * dashboard next, the same event marks them as landed.
- */
-it('flags the workspace created event when the wizard finishes', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Tracked Corp',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    expect(session()->get('fathom.track_workspace_created'))->toBeTrue();
-});
-
-/**
- * Copying the invite link pre-creates the workspace so the URL can exist. A
- * user who stops there has not finished onboarding, and counting them would
- * inflate the very conversion this event measures.
- */
-it('does not flag the workspace created event when the invite link only pre-creates the workspace', function (): void {
-    $user = User::factory()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Abandoned Corp',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-        ])
-        ->callAction(TestAction::make('copyInviteLink')->schemaComponent());
-
-    expect(Team::query()->where('name', 'Abandoned Corp')->exists())->toBeTrue()
-        ->and(session()->has('fathom.track_workspace_created'))->toBeFalse();
-});
-
-/**
- * A second workspace is expansion, not acquisition. Its Fathom referrer is
- * whatever brought the user back that day, so crediting a channel with it
- * would be wrong, and counting it alongside first workspaces would push the
- * signup-to-workspace rate past 100%.
- */
-it('does not flag the workspace created event for an additional workspace', function (): void {
-    $user = User::factory()->withPersonalTeam()->create();
-
-    $this->actingAs($user);
-
-    livewire(CreateTeam::class)
-        ->fillForm([
-            'name' => 'Second Corp',
-            'onboarding_use_case' => OnboardingUseCase::Other->value,
-        ])
-        ->call('register')
-        ->assertHasNoFormErrors();
-
-    expect(Team::query()->where('name', 'Second Corp')->exists())->toBeTrue()
-        ->and(session()->has('fathom.track_workspace_created'))->toBeFalse();
-});
